refactor for support all types read from hdf5 file
